improve caching for QMK if ci.nix has been built
It's easy to end up with multiple copies of the same source in /nix/store from
doing too much string interpolation.  I found a spot where this was happening,
which was causing a cache miss.  With this small change, ci.nix and the flash
script get to the same derivation for the keymaps.